move transactor client to transactor module
